requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
ewmh==0.1.6
idna==3.7
pulsectl==24.4.0
pyee==11.1.0
python-xlib==0.33
requests==2.32.3
six==1.16.0
touchportal-api==1.7.10
typing_extensions==4.12.2
urllib3==2.2.2
